Changes to build and run Titan modules and applications with stock Lua (without changes needed to `luaconf.h`)

+++ spec/coder_spec.lua
@@ -6,26 +6,34 @@ local util = require 'titan-compiler.util'
 local pretty = require 'titan-compiler.pretty'
 local driver = require 'titan-compiler.driver'
 
+local verbose = false
+
 local function parse(code)
     return parser.parse("(parser_spec)", code)
 end
 
-local function generate_modules(modules, main)
+local function generate_modules(modules, main, forapp)
     types.registry = {}
     local imported = {}
     local loader = driver.tableloader(modules, imported)
     local type, err = checker.checkimport(main, loader)
     if not type then return nil, err end
     if #err ~= 0 then return nil, table.concat(err, "\n") end
+    local ofiles = {}
     for name, mod in pairs(imported) do
-        local ok, err = driver.compile_module(name, mod)
+        local ok, err = driver.compile_module(name, mod, nil, forapp, verbose)
+        table.insert(ofiles, name .. ".o")
+        if not ok then return nil, err end
+    end
+    if forapp then
+        local ok, err = driver.compile_program(main, ofiles, nil, verbose)
         if not ok then return nil, err end
     end
     return true
 end
 
 local function call(modname, code)
-    local cmd = string.format("lua/src/lua -l %s -e \"print(pcall(function () %s end))\"",
+    local cmd = string.format("lua-5.3.4/src/lua -l %s -e \"print(pcall(function () %s end))\"",
         modname, code)
     local f = io.popen(cmd)
     local out = f:read()
@@ -46,7 +54,7 @@ local function run_coder(titan_code, lua_test, errmsg)
     assert.truthy(ast, err)
     local ok, err = checker.check("test", ast, titan_code, "test.titan")
     assert.equal(0, #err, table.concat(err, "\n"))
-    local ok, err = driver.compile("test", ast)
+    local ok, err = driver.compile("test", ast, nil, nil, verbose)
     assert.truthy(ok, err)
     local ok, err = call("test", lua_test)
     if errmsg then
@@ -57,10 +65,46 @@ local function run_coder(titan_code, lua_test, errmsg)
     end
 end
 
+local function call_app(appname, ...)
+    local cmd = table.concat({ appname, ... }, " ")
+    local f = io.popen(cmd)
+    local out = f:read()
+    local ok, err, status = f:close()
+    os.remove(appname)
+    return ok, err, status, out
+end
+
+local function run_coder_app(titan_code, main, estatus, eout)
+    types.registry = {}
+    local code = string.format([[
+        %s
+        function main(args: {string}): integer
+            %s
+        end
+    ]], titan_code, main)
+    local ast, err = parse(code)
+    assert.truthy(ast, err)
+    local ok, err = checker.check("test", ast, code, "test.titan")
+    assert.equal(0, #err, table.concat(err, "\n"))
+    local ok, err = driver.compile("test", ast, nil, nil, true, verbose)
+    assert.truthy(ok, err)
+    local ok, err = driver.compile_program("test", { "test.o" }, nil, verbose)
+    assert.truthy(ok, err)
+    local ok, err, status, output = call_app("./test")
+    assert.truthy(ok, err)
+    assert.equal(estatus, status)
+    if eout then
+        assert.match(eout, output)
+    end
+end
+
 describe("Titan code generator", function()
     after_each(function ()
+        os.execute("rm -f *.o")
         os.execute("rm -f *.so")
-        os.execute("rm -f *.c")
+        if not verbose then
+            os.execute("rm -f *.c")
+        end
     end)
 
     it("deletes array element", function()
@@ -2164,6 +2208,47 @@ describe("Titan code generator", function()
 
     end)
 
+    describe("#apps", function()
+        it("compiles and runs a titan application", function ()
+            run_coder_app([[
+                local function fn(m: { string: integer }): integer
+                    return m['foo']
+                end
+            ]], [[
+                local m = { ['foo'] = 0 }
+                return fn(m)
+            ]], 0)
+        end)
+
+        it("correctly uses module function in application", function ()
+            local modules = {
+                foo = [[
+                    function a(): integer
+                        return 42
+                    end
+                ]],
+                bar = [[
+                    local foo = import "foo"
+                    function bar(): integer
+                        return foo.a()
+                    end
+                    function main(args: {string}): integer
+                        if bar() == 42 then
+                            return 0
+                        else
+                            return 1
+                        end
+                    end
+                ]]
+            }
+            local ok, err = generate_modules(modules, "bar", true)
+            assert.truthy(ok, err)
+            local ok, err, status = call_app("./bar")
+            assert.truthy(ok, err)
+            assert.equal(0, status)
+        end)
+
+    end)
 end)
